% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcrs-package.R
\docType{package}
\name{wcrs-package}
\alias{wcrs}
\alias{wcrs-package}
\title{wcrs: Weighted Competing Risks analysis}
\description{
The package provides a series of tools to
    implement a Weighted Competing Risks analysis in a unified
    framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/wcrs}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/wcrs/issues}
}

}
\author{
\strong{Maintainer}: Daniele Bottigliengo \email{daniele.bottigliengo@phd.unipd.it} (\href{https://orcid.org/0000-0002-8802-3834}{ORCID})

}
\keyword{internal}
