% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cif.R
\name{weighted_cif}
\alias{weighted_cif}
\title{Estimate weighted cumulative incidence functions}
\usage{
weighted_cif()
}
\value{

}
\description{
Estimate weighted cumulative incidence functions
}
\examples{

}
